benchmarking/nursery/nas_plm/requirements.txt
tqdm
scipy
numpy
botorch==0.7.2
protobuf==3.20.*
pandas
datasets
scikit-learn
# torch
# tokenizers
tensorboardx
evaluate
transformers
# deepspeed>=0.8.2
sortedcontainers
# git+https://github.com/huggingface/accelerate.git
cupy-cuda11x  # for pruning
botorch==0.7.2
autograd>=1.3
pre-commit
